+++ cpskin/core/behaviors/indexview.py
@@ -15,6 +15,7 @@ class ICpskinIndexViewSettings(model.Schema):
         label=_(
             u"Index view"),
         fields=(
+            'collection_image_scale',
             'slider_image_scale',
             'carousel_image_scale',
             'link_text',
@@ -35,6 +36,14 @@ class ICpskinIndexViewSettings(model.Schema):
         description=_(u'Please select which fields should be visible.'),
         required=True,
         default='carousel',
+        vocabulary=u'plone.app.vocabularies.ImagesScales'
+    )
+
+    collection_image_scale = schema.Choice(
+        title=_(u"Which image scale use for collections of index view"),
+        description=_(u'Please select which fields should be visible.'),
+        required=False,
+        default='mini',
         vocabulary=u'plone.app.vocabularies.ImagesScales'
     )